TransactionCreateDto->TransactionDto org.postgresql.util.PSQLException: ПОМИЛКА: null значення в стовпці "receiver_account_iban" відношення "transactions" порушує not-null обмеження   Detail: Помилковий рядок містить (10, 2, null, null, null, UAH).

+++ src/main/java/org/simpleopenbanking/dto/TransactionDto.java
@@ -3,11 +3,10 @@ package org.simpleopenbanking.dto;
 import org.simpleopenbanking.enums.CurrencyType;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
  * DTO for {@link org.simpleopenbanking.entity.Transaction}
  */
-public record TransactionDto(long Id, long senderAccountIban, long receiverAccountIban, BigDecimal amount, CurrencyType currencyType, LocalDateTime timestamp) implements Serializable {
+public record TransactionDto(Long Id, Long senderAccountIban, Long receiverAccountIban, Long amount, CurrencyType currencyType, LocalDateTime timestamp) implements Serializable {
   }
